feat: used moved data arrays to index.js
