+++ first-project/src/App.js
@@ -16,15 +16,14 @@ const App = (props) => {
         <Header />
         <Navbar />
         <div className='app-wrapper-content'>
-          <Route path='/profile' component={Profile} />
-          <Route path='/dialogs' component={Dialogs} />
-          <Route path='/news' component={News} />
-          <Route path='/music' component={Music} />
-          <Route path='/settings' component={Settings} />
+          <Route path='/profile' render={ () => <Profile posts={props.posts}/>} />
+          <Route path='/dialogs' render={ () => <Dialogs dialogs={props.dialogs} messages={props.messages}/>} />
+          <Route path='/news' render={ () => <News />}/>
+          <Route path='/music' render={ () => <Music />} />
+          <Route path='/settings' render={ () => <Settings />} />
         </div>
       </div>
     </BrowserRouter>
-    
   );
 }
