+++ first-project/src/components/Dialogs/Dialogs.module.css
@@ -10,12 +10,7 @@
     background-color: #fff;
 }
 
-
 .messages {
     padding: 10px;
     background-color: #fff;
-}
-
-.message {
-
 }
