+++ first-project/src/components/Dialogs/Message/Message.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import classes from './Message.module.css';
-import {NavLink} from "react-router-dom";
 
 const Message = (props) => {
   return (
