+++ first-project/src/components/Profile/MyPosts/MyPosts.jsx
@@ -2,13 +2,9 @@ import React from 'react';
 import classes from './MyPosts.module.css';
 import Post from './Post/Post';
 
-const MyPosts = () => {
-  let posts = [
-    {id: 1, message: 'We bomb the ReactJS at the rate of Dimych!', likeCount: '15'},
-    {id: 2, message: 'Hello! I am learning ReactJS!', likeCount: '20'}
-  ];
-
-  let postsElements = posts.map( p => <Post message={p.message} likeCount={p.likeCount} /> );
+const MyPosts = (props) => {
+  let postsElements = props.items
+    .map( (p, index) => <Post message={p.message} likeCount={p.likeCount}  key={index}/> );
 
   return (
     <div className={classes.posts}>
